feat: update student info component
